Networth item handling rewrite
