+++ build.gradle
@@ -102,8 +102,21 @@ repositories {
 		}
 	}
 
-	maven { url "https://maven.azureaaron.net/releases" }
-	maven { url "https://maven.azureaaron.net/snapshots" }
+	maven {
+		url "https://maven.azureaaron.net/releases"
+
+		content {
+			includeGroup "net.azureaaron"
+		}
+	}
+
+	maven {
+		url "https://maven.azureaaron.net/snapshots"
+
+		content {
+			includeGroup "net.azureaaron"
+		}
+	}
 }
 
 dependencies {
@@ -148,7 +161,7 @@ dependencies {
 	// NEU RepoParser
 	include implementation("moe.nea:neurepoparser:${project.repoparser_version}")
 
-	// Networth Calculator
+	// Networth Calculator (https://github.com/AzureAaron/networth-calculator)
 	include implementation("net.azureaaron:networth-calculator:${project.networth_calculator_version}")
 
 	// JGit used pull data from the NEU item repo
