+++ gradle.properties
@@ -30,8 +30,8 @@ mcdev_annotations_version = 2.1.0
 occlusionculling_version = 0.0.8-SNAPSHOT
 ## neu repoparser (https://repo.nea.moe/#/releases/moe/nea/neurepoparser/)
 repoparser_version = 1.5.0
-## Networth Calculator
-networth_calculator_version = 1.0.0-alpha.2
+## Networth Calculator (https://github.com/AzureAaron/networth-calculator)
+networth_calculator_version = 1.0.0-alpha.3
 
 # Other Libraries
 ## JGit (https://mvnrepository.com/artifact/org.eclipse.jgit/org.eclipse.jgit)
