+++ src/main/java/de/hysky/skyblocker/utils/networth/NetworthCalculator.java
@@ -1,5 +1,7 @@
 package de.hysky.skyblocker.utils.networth;
 
+import com.mojang.serialization.Dynamic;
+
 import de.hysky.skyblocker.utils.ItemUtils;
 import net.azureaaron.networth.ItemCalculator;
 import net.azureaaron.networth.NetworthResult;
@@ -7,12 +9,17 @@ import net.azureaaron.networth.item.SkyblockItemStack;
 import net.minecraft.component.DataComponentTypes;
 import net.minecraft.component.type.NbtComponent;
 import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.NbtCompound;
+import net.minecraft.nbt.NbtElement;
+import net.minecraft.nbt.NbtOps;
 
 public class NetworthCalculator {
 
 	public static NetworthResult getItemNetworth(ItemStack stack) {
 		String itemId = ItemUtils.getItemId(stack);
-		SkyblockItemStack skyblockItemStack = SkyblockItemStack.of(itemId, stack.getCount(), SkyblockItemMetadata.of(stack.getOrDefault(DataComponentTypes.CUSTOM_DATA, NbtComponent.DEFAULT).copyNbt(), itemId));
+		NbtCompound customData = stack.getOrDefault(DataComponentTypes.CUSTOM_DATA, NbtComponent.DEFAULT).copyNbt();
+		Dynamic<NbtElement> customDataDynamic = new Dynamic<>(NbtOps.INSTANCE, customData);
+		SkyblockItemStack skyblockItemStack = SkyblockItemStack.of(itemId, stack.getCount(), customDataDynamic, SkyblockItemMetadataRetriever.of(customData, itemId));
 
 		return ItemCalculator.calculate(skyblockItemStack, NetworthDataSuppliers::getPrice, NetworthDataSuppliers.getSkyblockItemData());
 	}
