+++ src/main/java/de/hysky/skyblocker/utils/networth/NetworthDataSuppliers.java
@@ -1,7 +1,5 @@
 package de.hysky.skyblocker.utils.networth;
 
-import java.util.List;
-
 import org.slf4j.Logger;
 
 import com.google.gson.JsonArray;
@@ -9,21 +7,23 @@ import com.mojang.logging.LogUtils;
 import com.mojang.serialization.JsonOps;
 
 import de.hysky.skyblocker.utils.ItemUtils;
+import it.unimi.dsi.fastutil.objects.Object2ObjectMap;
+import it.unimi.dsi.fastutil.objects.Object2ObjectMaps;
 import net.azureaaron.networth.data.SkyblockItemData;
 
 public class NetworthDataSuppliers {
 	private static final Logger LOGGER = LogUtils.getLogger();
-	private static List<SkyblockItemData> itemData = List.of();
+	private static Object2ObjectMap<String, SkyblockItemData>  itemData = Object2ObjectMaps.emptyMap();
 
 	public static void updateSkyblockItemData(JsonArray items) {
 		try {
-			itemData = SkyblockItemData.LIST_CODEC.parse(JsonOps.INSTANCE, items).getOrThrow();
+			itemData = SkyblockItemData.MAP_CODEC.parse(JsonOps.INSTANCE, items).getOrThrow();
 		} catch (Exception e) {
 			LOGGER.error("[Skyblocker Networth Data Supplier] Failed to parse items data!", e);
 		}
 	}
 
-	static List<SkyblockItemData> getSkyblockItemData() {
+	static Object2ObjectMap<String, SkyblockItemData> getSkyblockItemData() {
 		return itemData;
 	}
